Code tidy, change parsers, add simple datatypes
This is a bit of a mish-mash commit. Changes xmlDecodeRecord and
xmlEncodeRecord by appending 'Type', just to make it clearer.
Parsers now operate on strings as their input, returning both the
result AND the unparsed string. This is probably less efficient, but
it does mean that the framework is much simpler to use. Finally we add
the encoders and decoders for cite keys and record types.